baekjoon bronze-4 15700

+++ baekjoon/bronze 4/15700/src/Main.java
@@ -2,8 +2,8 @@ import java.util.Scanner;
 public class Main {
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
-		int N = sc.nextInt();
-		int M = sc.nextInt();
+		long N = sc.nextLong();
+		long M = sc.nextLong();
 		
 		System.out.println((N*M)/2);
 	}
